backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
Werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
Pillow==10.0.1
boto3==1.29.7
twilio==8.8.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2